requirements.txt
asgiref==3.7.2
Django==5.0.3
mysql-replication==1.0.7
packaging==23.2
PyMySQL==1.1.0
sqlparse==0.4.4
